Link to homepage
Feature: Whitelisting for a set of users. If at least one user is selected the stream is only viewable for this viewer. If the selection is empty the stream is public for all registered users.

This feature needs the nginx auth module (see docker environmet for an example)

Added auth and token logic to viewcount and thumbs logic, now unregistered and non-whitelisted users are not able to see thumbnails and view-stats

Added script to generate nginx self signed certs
Added script to build docker container for ovenauth
Added script to publish local frontend

Added docker-compose file to set up test environment locally